+++ src/plugins/object.ts
@@ -9,16 +9,19 @@ import mongoose from "mongoose";
  */
 export function flattenObject(obj: Record<string, any>): Record<string, any> {
 	const flattenKeys = (_obj: Record<string, any>, prefix = ""): Record<string, any> => {
-		return Object.keys(_obj).reduce((acc, key) => {
-			const newKey = prefix ? `${prefix}.${key}` : key;
-			const value = _obj[key];
-			if (typeof value === "object" && value !== null && !(value instanceof mongoose.Types.ObjectId) && !(value instanceof Date)) {
-				Object.assign(acc, flattenKeys(value, newKey));
-			} else {
-				acc[newKey] = value;
-			}
-			return acc;
-		}, {} as Record<string, any>);
+		return Object.keys(_obj).reduce(
+			(acc, key) => {
+				const newKey = prefix ? `${prefix}.${key}` : key;
+				const value = _obj[key];
+				if (typeof value === "object" && value !== null && !(value instanceof mongoose.Types.ObjectId) && !(value instanceof Date)) {
+					Object.assign(acc, flattenKeys(value, newKey));
+				} else {
+					acc[newKey] = value;
+				}
+				return acc;
+			},
+			{} as Record<string, any>
+		);
 	};
 
 	const flattenedObj = flattenKeys(obj);