+++ src/plugins/utils.ts
@@ -936,9 +936,12 @@ export const resolveFilePath = (fileNamePrefix: string, options: ResolveApplicat
 	filePath = path.resolve(targetDirectory, `deployment/${fileNamePrefix}`);
 	if (fs.existsSync(filePath)) return filePath;
 
+	filePath = path.resolve(targetDirectory, fileNamePrefix);
+	if (fs.existsSync(filePath)) return filePath;
+
 	if (!ignoreIfNotExisted) {
-		const message = `Missing "${targetDirectory}/${fileNamePrefix}" file, please create one.`;
-		logError(message);
+		const message = `Missing "./${fileNamePrefix}" file, are you in the project directory?`;
+		throw new Error(message);
 	}
 	return;
 };